+++ lib/rubocop/cop/performance/map_compact.rb
@@ -66,8 +66,8 @@ module RuboCop
           chained_method = compact_node.parent
           compact_method_range = compact_node.loc.selector
 
-          if compact_node.multiline? &&
-             chained_method && !invoke_method_after_map_compact_on_same_line?(compact_node, chained_method)
+          if compact_node.multiline? && chained_method&.loc.respond_to?(:selector) &&
+             !invoke_method_after_map_compact_on_same_line?(compact_node, chained_method)
             range_by_whole_lines(compact_method_range, include_final_newline: true)
           else
             compact_method_range